+++ packages/html/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @web-widget/html
 
+## 1.18.0
+
+### Patch Changes
+
+- @web-widget/helpers@1.18.0
+
+## 1.17.0
+
+### Patch Changes
+
+- @web-widget/helpers@1.17.0
+
+## 1.16.1
+
+### Patch Changes
+
+- @web-widget/helpers@1.16.1
+
+## 1.16.0
+
+### Patch Changes
+
+- @web-widget/helpers@1.16.0
+
 ## 1.15.1
 
 ### Patch Changes