+++ packages/lifecycle-cache/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @web-widget/lifecycle-cache
 
+## 1.18.0
+
+### Patch Changes
+
+- @web-widget/context@1.18.0
+
+## 1.17.0
+
+### Patch Changes
+
+- @web-widget/context@1.17.0
+
+## 1.16.1
+
+### Patch Changes
+
+- @web-widget/context@1.16.1
+
+## 1.16.0
+
+### Patch Changes
+
+- @web-widget/context@1.16.0
+
 ## 1.15.1
 
 ### Patch Changes