+++ packages/react/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @web-widget/react
 
+## 1.18.0
+
+### Patch Changes
+
+- @web-widget/vite-plugin@1.18.0
+- @web-widget/helpers@1.18.0
+- @web-widget/web-widget@1.18.0
+
+## 1.17.0
+
+### Patch Changes
+
+- @web-widget/helpers@1.17.0
+- @web-widget/vite-plugin@1.17.0
+- @web-widget/web-widget@1.17.0
+
+## 1.16.1
+
+### Patch Changes
+
+- Updated dependencies [6d3900c]
+  - @web-widget/web-widget@1.16.1
+  - @web-widget/vite-plugin@1.16.1
+  - @web-widget/helpers@1.16.1
+
+## 1.16.0
+
+### Patch Changes
+
+- @web-widget/helpers@1.16.0
+- @web-widget/vite-plugin@1.16.0
+- @web-widget/web-widget@1.16.0
+
 ## 1.15.1
 
 ### Patch Changes